+++ widgets/Login.php
@@ -1,0 +1,46 @@
+<?php
+
+/*
+ * This file is part of the Dektrium project.
+ *
+ * (c) Dektrium project <http://github.com/dektrium>
+ *
+ * For the full copyright and license information, please view the LICENSE.md
+ * file that was distributed with this source code.
+ */
+
+namespace dektrium\user\widgets;
+
+use dektrium\user\helpers\ModuleTrait;
+use yii\base\Widget;
+
+/**
+ * @author Dmitry Erofeev <dmeroff@gmail.com>
+ */
+class Login extends Widget
+{
+    use ModuleTrait;
+
+    /**
+     * @var bool
+     */
+    public $validate = true;
+
+    /**
+     * @inheritdoc
+     */
+    public function run()
+    {
+        $model  = $this->module->manager->createLoginForm();
+        $action = $this->validate ? null : ['/user/security/login'];
+
+        if ($this->validate && $model->load(\Yii::$app->getRequest()->post()) && $model->login()) {
+            return \Yii::$app->response->redirect(\Yii::$app->user->returnUrl);
+        }
+
+        return $this->render('login', [
+            'model'  => $model,
+            'action' => $action
+        ]);
+    }
+}