+++ widgets/views/login.php
@@ -1,0 +1,47 @@
+<?php
+
+/*
+ * This file is part of the Dektrium project.
+ *
+ * (c) Dektrium project <http://github.com/dektrium>
+ *
+ * For the full copyright and license information, please view the LICENSE.md
+ * file that was distributed with this source code.
+ */
+
+use yii\widgets\ActiveForm;
+use yii\helpers\Html;
+
+/**
+ * @var yii\web\View $this
+ * @var yii\widgets\ActiveForm $form
+ * @var dektrium\user\models\LoginForm $model
+ * @var string $action
+ */
+?>
+
+<?php if (Yii::$app->user->isGuest): ?>
+
+    <?php $form = ActiveForm::begin([
+        'id' => 'login-widget-form',
+        'fieldConfig' => [
+            'template' => "{input}\n{error}"
+        ],
+        'action' => $action
+    ]) ?>
+
+    <?= $form->field($model, 'login')->textInput(['placeholder' => 'Login']) ?>
+
+    <?= $form->field($model, 'password')->passwordInput(['placeholder' => 'Password']) ?>
+
+    <?= $form->field($model, 'rememberMe')->checkbox() ?>
+
+    <?= Html::submitButton(Yii::t('user', 'Sign in'), ['class' => 'btn btn-primary btn-block']) ?>
+
+    <?php ActiveForm::end(); ?>
+
+<?php else: ?>
+
+    <?= Html::a(Yii::t('user', 'Logout'), ['/user/security/logout'], ['class' => 'btn btn-danger btn-block', 'data-method' => 'post']) ?>
+
+<?php endif ?>